refactor: Restructure shared folder
