+++ src/shared/Components/Common/Switch.tsx
@@ -1,5 +1,5 @@
 import * as SwitchPrimitives from '@radix-ui/react-switch';
-import { cn } from 'lib/utils';
+import { cn } from 'shared/lib/utils';
 import * as React from 'react';
 
 interface SwitchOption<T> {
