+++ src/shared/Components/ui/avatar.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as AvatarPrimitive from "@radix-ui/react-avatar"
 
-import { cn } from "@/lib/utils"
+import { cn } from "shared/lib/utils"
 
 function Avatar({
   className,
