+++ src/shared/Components/ui/dropdown-menu.tsx
@@ -3,7 +3,7 @@ import * as React from "react"
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu"
 import { Check, ChevronRight, Circle } from "lucide-react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "shared/lib/utils"
 
 const DropdownMenu = DropdownMenuPrimitive.Root
