+++ src/shared/Components/ui/switch.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as SwitchPrimitive from "@radix-ui/react-switch"
 
-import { cn } from "@/lib/utils"
+import { cn } from "shared/lib/utils"
 
 function Switch({
   className,
